modularize
