+++ graph_box_umbrella/apps/web_client/src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import './App.scss';
 import BasicColumnChart from './components/ColumnCharts/Basic/index';
 import ColumnWithLabels from './components/ColumnCharts/ColumnWithLabels/index';
-import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
+import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import Menu from './menu';
 
 function App() {
