--- graph_box_umbrella/apps/web_client/src/components/ColumnCharts/ColumnWithLabels/classes.js
@@ -1,33 +1,0 @@
-export const styles = theme => ({
-  textField: {
-    marginLeft: 8,
-    marginRight: 8
-  },
-  button: {
-    margin: 8
-  },
-  editTextField: {
-    margin: 10,
-    marginBottom: 24
-  },
-  valuesTextField: {
-    margin: 10
-  },
-  fieldLabel: {
-    display: "block",
-    fontSize: 20,
-    padding: "0 10px 0 10px"
-  },
-  addDataButton: {
-    minWidth: 40,
-    minheight: 40
-  },
-  chartControl: {
-    display: "flex"
-  },
-  chart: {
-    paddingLeft: 343,
-    height: "100%",
-    width: "calc(100% - 335px)"
-  }
-})
