+++ graph_box_umbrella/apps/web_client/src/components/ColumnCharts/ColumnWithLabels/index.js
@@ -1,10 +1,12 @@
 import React, { Component } from "react";
 import Chart from 'react-apexcharts';
-import { LoopSeries, LoopCategories } from './loops';
+import { LoopSeries, LoopCategories } from '../loops';
+import { sharedSeriesConfig, sharedOptionsConfig } from '../sharedConfig';
+import { addNewSeries, editCategories, editTitle, editSeriesTitle, editSeriesValues, addNewCategory } from '../sharedEventHandlers';
 
 //@material-ui styling
 import { withStyles } from '@material-ui/core/styles';
-import { styles } from './classes';
+import { styles } from '../classes';
 
 
 //@material-ui components
@@ -17,24 +19,18 @@ import Fab from '@material-ui/core/Fab';
 import Typography from '@material-ui/core/Typography';
 import TextField from '@material-ui/core/TextField';
 
-class ColumnWithLabels extends React.Component {
+class ColumnWithLabels extends Component {
   constructor() {
     super();
 
     this.state = {
-      series: [{
-        name: 'Inflation',
-        data: [2.3, 3.1, 4.0, 10.1, 4.0, 3.6, 3.2, 2.3, 1.4, 0.8, 0.5, 0.2]
-      }],
+      ...sharedSeriesConfig(),
       options: {
-        chart: {
-          height: 450,
-          type: 'bar',
-        },
+        ...sharedOptionsConfig(),
         plotOptions: {
           bar: {
             dataLabels: {
-              position: 'top', // top, center, bottom
+              position: 'top'
             },
           }
         },
@@ -49,9 +45,8 @@ class ColumnWithLabels extends React.Component {
             colors: ["#304758"]
           }
         },
-
         xaxis: {
-          categories: ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"],
+          categories: [],
           position: 'top',
           axisBorder: {
             show: false
@@ -88,120 +83,16 @@ class ColumnWithLabels extends React.Component {
               return val + "%";
             }
           }
-
-        },
-        title: {
-          text: 'Monthly Inflation in Argentina, 2002',
-          floating: true,
-          margin: 50,
-          align: 'center',
-          style: {
-            color: '#444'
-          }
         }
-      },
+      }
     };
 
-    this.editCategories = this.editCategories.bind(this);
-    this.editSeriesTitle = this.editSeriesTitle.bind(this);
-    this.editSeriesValues = this.editSeriesValues.bind(this);
-  }
-
-  editCategories(e, categoryIndex) {
-    const { value } = e.target;
-    if (value !== "") {
-      const categoriesCopy = [...this.state.options.xaxis.categories]
-
-      categoriesCopy[categoryIndex] = value
-
-      this.setState({
-        options: {
-          ...this.state.options,
-          xaxis: {
-            categories: categoriesCopy
-          }
-        }
-      });
-    }
-  }
-
-  editTitle(e) {
-    const { value } = e.target;
-
-    this.setState({
-      options: {
-        ...this.state.options,
-        title: {
-          ...this.state.options.title,
-          text: value
-        }
-      }
-    })
-  }
-
-  editSeriesTitle(e, seriesName) {
-    const { value } = e.target
-    if (value !== "") {
-      const seriesCopy = [...this.state.series]
-      const index = seriesCopy.findIndex((obj) => obj.name === seriesName)
-
-      seriesCopy[index] = { ...seriesCopy[index], name: value }
-
-      this.setState({
-        series: seriesCopy
-      })
-    }
-  }
-
-  editSeriesValues(e, index, seriesName) {
-    const { value } = e.target
-
-    if (value !== "" && parseInt(value)) {
-      this.setState(prevState => ({
-        series: prevState.series.map((series) => {
-          if (series.name !== seriesName) {
-            return series
-          }
-
-          return {
-            ...series,
-            data: series.data.map((dataValue, dataIndex) => {
-              if (dataIndex !== index) {
-                return dataValue
-              }
-
-              return value
-            })
-          }
-        })
-      }))
-    }
-  }
-
-  addNewCategory() {
-    const { categories } = this.state.options.xaxis
-    this.setState({
-      series: this.state.series.map(({ data, name }) => {
-        return {
-          data: data.concat([0]),
-          name
-        }
-      }),
-      options: {
-        ...this.state.options,
-        xaxis: {
-          categories: [...categories, ...[""]]
-        }
-      }
-    })
-  }
-
-  addNewSeries(e) {
-    const data = this.state.options.xaxis.categories.map(() => 0)
-
-    this.setState({
-      series: [...this.state.series, { name: "", data }]
-    })
+    this.addNewSeries = addNewSeries.bind(this);
+    this.editCategories = editCategories.bind(this);
+    this.editTitle = editTitle.bind(this);
+    this.editSeriesTitle = editSeriesTitle.bind(this);
+    this.editSeriesValues = editSeriesValues.bind(this);
+    this.addNewCategory = addNewCategory.bind(this);
   }
 
   render() {
@@ -244,7 +135,6 @@ class ColumnWithLabels extends React.Component {
             </Fab>
           </ExpansionPanelDetails>
         </ExpansionPanel>
-
       </div>
     );
   }
