+++ graph_box_umbrella/apps/web_client/src/components/ColumnCharts/classes.js
