+++ graph_box_umbrella/apps/web_client/src/components/ColumnCharts/sharedConfig.js
@@ -1,0 +1,20 @@
+export const sharedSeriesConfig = () => ({
+  series: [{
+    name: '',
+    data: []
+  }]
+});
+
+export const sharedOptionsConfig = () => ({
+  chart: {
+    type: 'bar'
+  },
+  title: {
+    text: '',
+    margin: 50,
+    align: 'center',
+    style: {
+      color: '#444'
+    }
+  }
+});
